fix(v30.3): build fixes — resolve pre-existing compilation errors
- SafeBootController.h: add QtQml include for QML_ELEMENT
- SafeBootAutoRepair.cpp: add QThread include for msleep
- AssistantManager.cpp: fix SafeBootController API method names
- CMakeLists.txt: add app/safeboot to target_include_directories
- tests/cpp/CMakeLists.txt: add SafeBootAutoRepair to test library
- SpatialEnums.h: add RequestHuman to ActionType, Modified to SupervisorDecision
- SpatialMemory.h/.cpp: rename MemoryEntry to SpatialMemoryEntry (fix redefinition with AIMemoryManager)
- SpatialSupervisor.cpp: fix toMap() -> toVariantMap()
- SpatialCognitiveEngine.cpp: fix storeRisk() args, getSpatialExplanation return type
- SpatialSecurityEngine.h: remove orphan #endif with #pragma once
- SpatialSecurityEngine.cpp: fix update*State() to use QVariantMap
- VisionDetections.cpp, VisionMemory.cpp: fix std::min/max qsizetype/int ambiguity

+++ app/vision/VisionDetections.cpp
@@ -240,7 +240,7 @@ QVariantList VisionDetections::currentDetections() const
 QVariantList VisionDetections::recentEvents() const
 {
     QVariantList list;
-    int start = std::max(0, m_events.size() - 100);
+    int start = std::max(0, static_cast<int>(m_events.size()) - 100);
     for (int i = m_events.size() - 1; i >= start; --i)
         list.append(m_events[i].toVariant());
     return list;
